feat: keypair export/import and utils clean up
