+++ packages/ucan-storage/src/keypair.js
@@ -1,6 +1,6 @@
-import { base64Pad } from './rfc4648.js'
 import * as ed from '@noble/ed25519'
 import { publicKeyBytesToDid } from './did.js'
+import { base64Pad } from './encoding.js'
 
 export class KeyPair {
   /**
@@ -11,6 +11,23 @@ export class KeyPair {
     this.publicKey = publicKey
     this.privateKey = privateKey
     this.pubStr = undefined
+  }
+
+  static async create() {
+    const privateKeyRaw = ed.utils.randomPrivateKey()
+    const publicKey = await ed.getPublicKey(privateKeyRaw)
+
+    return new KeyPair(privateKeyRaw, publicKey)
+  }
+
+  /**
+   * Create keypair from exported private key
+   *
+   * @param {string} key
+   */
+  static async fromExportedKey(key) {
+    const privateKey = base64Pad.decode(key)
+    return new KeyPair(privateKey, await ed.getPublicKey(privateKey))
   }
 
   publicKeyStr() {
@@ -32,32 +49,7 @@ export class KeyPair {
     return ed.sign(msg, this.privateKey)
   }
 
-  static async create() {
-    const privateKeyRaw = ed.utils.randomPrivateKey()
-    const publicKey = await ed.getPublicKey(privateKeyRaw)
-
-    return new KeyPair(
-      //   concatKeys(privateKeyRaw, publicKey),
-      privateKeyRaw,
-      publicKey
-    )
+  export() {
+    return base64Pad.encode(this.privateKey)
   }
 }
-
-// /**
-//  * @param {Uint8Array} privateKeyRaw
-//  * @param {Uint8Array} publicKey
-//  */
-// function concatKeys(privateKeyRaw, publicKey) {
-//   const privateKey = new Uint8Array(64)
-//   for (let i = 0; i < 32; i++) {
-//     privateKey[i] = privateKeyRaw[i]
-//     privateKey[32 + i] = publicKey[i]
-//   }
-//   return privateKey
-// }
-
-// async function main() {
-//   await KeyPair.create()
-// }
-// main()
